refactor view and create custom classes for views
