+++ quavi/quavi/Controllers/MapViewController+SliderView.swift
@@ -8,28 +8,159 @@
 
 import UIKit
 
+extension MapViewController {
+    
+    func handleCollectionViewCellPressed(item: Int) {
+        if item == 0 {
+            currentSelectedCategory = Enums.categories.History.rawValue
+        } else if item == 1 {
+            currentSelectedCategory = Enums.categories.Art.rawValue
+        } else if item == 2 {
+            currentSelectedCategory = Enums.categories.Science.rawValue
+        } else if item == 3 {
+            currentSelectedCategory = Enums.categories.Religion.rawValue
+        } else if item == 4 {
+            currentSelectedCategory = Enums.categories.Yeet.rawValue
+        }
+    }
+    
+    func loadGestures() {
+        let swipeDown = UISwipeGestureRecognizer(target: self, action: #selector(respondToSwipeGesture(gesture:)))
+        swipeDown.direction = .down
+        self.sliderView.addGestureRecognizer(swipeDown)
+        
+        let swipeUp = UISwipeGestureRecognizer(target: self, action: #selector(respondToSwipeGesture(gesture:)))
+        swipeUp.direction = UISwipeGestureRecognizer.Direction.up
+        self.sliderView.addGestureRecognizer(swipeUp)
+        
+        let tap = UITapGestureRecognizer(target: self, action: #selector(respondToSwipeGesture(gesture:)))
+        self.chevronArrows.addGestureRecognizer(tap)
+    }
+    
+    func directionOfChevron(state: Enums.sliderViewStates) {
+        
+        switch state {
+        case .halfOpen:
+            self.chevronArrows.image = UIImage(systemName: "minus")
+        case .fullOpen:
+            self.chevronArrows.image = UIImage(systemName: "chevron.compact.down")
+        case .closed:
+            self.chevronArrows.image = UIImage(systemName: "chevron.compact.up")
+        }
+    }
+    
+    @objc func respondToSwipeGesture(gesture: UIGestureRecognizer) {
+        print(gesture)
+        
+        if let tapGesture = gesture as? UITapGestureRecognizer {
+            print("tapped")
+            switch tapGesture.numberOfTouches {
+            case 1:
+                sliderViewState = .halfOpen
+                
+                UIView.animate(withDuration: 1.0, delay: 0, usingSpringWithDamping: 0.80, initialSpringVelocity: 0, options: .curveEaseInOut, animations: { [weak self] in
+                    self?.setHalfOpenSliderViewConstraints()
+                    self?.directionOfChevron(state: .halfOpen)
+                    self?.view.layoutIfNeeded()
+                    self?.sliderView.alpha = 1.0
+                    self?.poiTableView.alpha = 1.0
+                    self?.categoriesCollectionView.alpha = 1.0
+                    }, completion: nil)
+            default:
+                print("dunno know")
+            }
+            
+        }
+        
+        if let swipeGesture = gesture as? UISwipeGestureRecognizer {
+            switch swipeGesture.direction {
+            case UISwipeGestureRecognizer.Direction.down:
+                print("Swiped down")
+                
+                UIView.animate(withDuration: 1.0, delay: 0, usingSpringWithDamping: 0.80, initialSpringVelocity: 0, options: .curveEaseInOut, animations: { [weak self] in
+                    
+                    switch self?.sliderViewState {
+                    case .fullOpen:
+                        self?.setHalfOpenSliderViewConstraints()
+                        self?.sliderViewState = .halfOpen
+                    case .halfOpen:
+                        self?.setClosedSliderViewConstraints()
+                        self?.sliderViewState = .closed
+                    case .closed:
+                        print("it's already closed")
+                    case .none:
+                        return
+                    }
+                    
+                    if self?.sliderViewState == .closed {
+                        self?.directionOfChevron(state: .closed)
+                    } else if self?.sliderViewState == .halfOpen {
+                        self?.directionOfChevron(state: .halfOpen)
+                    } else if self?.sliderViewState == .fullOpen {
+                        self?.directionOfChevron(state: .fullOpen)
+                    }
+                    self?.view.layoutIfNeeded()
+                    
+                    if self?.sliderViewState == .closed {
+                        self?.sliderView.alpha = 0.5
+                        self?.poiTableView.alpha = 0
+                        self?.categoriesCollectionView.alpha = 0
+                    }
+                    
+                    }, completion: nil)
+                
+            case UISwipeGestureRecognizer.Direction.up:
+                UIView.animate(withDuration: 1.0, delay: 0, usingSpringWithDamping: 0.80, initialSpringVelocity: 0, options: .curveEaseInOut, animations: { [weak self] in
+                    
+                    switch self?.sliderViewState {
+                    case .fullOpen:
+                        print("it's fully opened")
+                    case .halfOpen:
+                        self?.setFullOpenSliderViewConstraints()
+                        self?.sliderViewState = .fullOpen
+                    case .closed:
+                        self?.setHalfOpenSliderViewConstraints()
+                        self?.sliderViewState = .halfOpen
+                    case .none:
+                        return
+                    }
+                    
+                    if self?.sliderViewState == .closed {
+                        self?.directionOfChevron(state: .closed)
+                    } else if self?.sliderViewState == .halfOpen {
+                        self?.directionOfChevron(state: .halfOpen)
+                    } else if self?.sliderViewState == .fullOpen {
+                        self?.directionOfChevron(state: .fullOpen)
+                    }
+                    self?.view.layoutIfNeeded()
+                    
+                    self?.sliderView.alpha = 1.0
+                    self?.poiTableView.alpha = 1.0
+                    self?.categoriesCollectionView.alpha = 1.0
+                    }, completion: nil)
+                
+            default:
+                break
+            }
+        }
+    }
+}
+
 //MARK: -EXT. TABLEVIEW DELEGATE & DATASOURCE
 extension MapViewController: UITableViewDelegate, UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        
-        if sampleData[section].isCellExpanded == false {
-            return 0
-        } else {
-            return 1
-        }
-        
+        if sampleData[section].isCellExpanded == false { return 0 } else { return 1 }
     }
     
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return sampleData.count
-    }
+    func numberOfSections(in tableView: UITableView) -> Int { return sampleData.count }
     
     func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
         
         if currentSelectedCategory == Enums.categories.History.rawValue {
             let view = UIView(frame: CGRect(x: 0, y: 0, width: self.view.frame.width, height: 50))
             
+            //TO-DO: SEPERATE INTO IT'S OWN FILE
             let button = UIButton(frame: CGRect(x: 0, y: 0, width: self.view.frame.width, height: 50))
             button.setTitle(sampleData[section].name, for: .normal)
             button.backgroundColor = .yellow
@@ -45,9 +176,7 @@ extension MapViewController: UITableViewDelegate, UITableViewDataSource {
         return nil
     }
     
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 50
-    }
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat { return 50 }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let stop = sampleData[indexPath.section]
@@ -68,7 +197,7 @@ extension MapViewController: UITableViewDelegate, UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         return UITableView.automaticDimension
+        
     }
-    
     
 }
